Move request response to version dir
