+++ common/ip-packet-requests/src/v6/mod.rs
@@ -1,0 +1,2 @@
+pub mod request;
+pub mod response;
